fix(test): keep advisory RLIMIT test off Windows
