+++ internal/advisory/sync_write_error_windows_test.go
@@ -1,0 +1,9 @@
+//go:build windows
+
+package advisory
+
+import "testing"
+
+func TestDownloadSnapshotWriteError(t *testing.T) {
+	t.Skip("RLIMIT_FSIZE is unavailable on Windows")
+}
